[connectivity_plus_web + linux] add @dart=2.9 (temporarily)
