+++ packages/connectivity_plus_linux/test/connectivity_plus_linux_test.dart
@@ -1,10 +1,10 @@
+//@dart=2.9
+
 import 'package:connectivity_plus_linux/src/connectivity_real.dart';
 import 'package:connectivity_plus_linux/src/network_manager.dart';
 import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-
-class MockNetworkManager extends Mock implements NetworkManager {}
 
 void main() {
   test('wireless', () async {
@@ -45,3 +45,5 @@ void main() {
         emitsInOrder([ConnectivityResult.wifi, ConnectivityResult.none]));
   });
 }
+
+class MockNetworkManager extends Mock implements NetworkManager {}
