+++ packages/connectivity_plus_web/example/test_driver/README.md
@@ -1,0 +1,3 @@
+## Running tests
+- `chromedriver -port=4444`
+- `flutter drive --target=test_driver/connectivity_plus_example.dart --show-web-server-device -d web-server`
