+++ packages/connectivity_plus_web/example/test_driver/connectivity_plus_example_test.dart
@@ -1,0 +1,5 @@
+//@dart=2.9
+
+import 'package:integration_test/integration_test_driver.dart';
+
+Future<void> main() async => integrationDriver();
