+++ packages/connectivity_plus_web/example/test_driver/src/connectivity_mocks.dart
@@ -1,3 +1,5 @@
+//@dart=2.9
+
 import 'dart:html';
 
 import 'package:mockito/mockito.dart';
